requirements.txt
fastapi
uvicorn
langchain-groq
python-dotenv
pydantic
requests